Chore(filter): export types properly
